requirements.txt
gspread==6.0.2
google-auth==2.27.0
selenium==4.17.2
